requirements.txt

streamlit==1.32.2
PyPDF2==3.0.1
pillow==10.4.0
pytesseract==0.3.10
requests==2.32.3
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0

# aiohttp==3.9.5
# aiosignal==1.3.1
# altair==5.3.0
# annotated-types==0.7.0
# anyio==4.4.0
# attrs==23.2.0
# blinker==1.8.2
# cachetools==5.4.0
# certifi==2024.7.4
# charset-normalizer==3.3.2
# click==8.1.7
# distro==1.9.0
# frozenlist==1.4.1
# gitdb==4.0.11
# GitPython==3.1.43
# google-ai-generativelanguage==0.6.6
# google-api-core==2.19.1
# google-api-python-client==2.138.0
# google-auth==2.32.0
# google-auth-httplib2==0.2.0
# google-generativeai==0.7.2
# googleapis-common-protos==1.63.2
# grpcio==1.65.1
# grpcio-status==1.62.2
# h11==0.14.0
# httpcore==1.0.5
# httplib2==0.22.0
# httpx==0.27.0
# idna==3.7
# Jinja2==3.1.4
# jsonschema==4.23.0
# jsonschema-specifications==2023.12.1
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# mdurl==0.1.2
# multidict==6.0.5
# numpy==1.26.4
# openai==0.28.0
# packaging==23.2
# pandas==2.2.2
# pillow==10.4.0
# proto-plus==1.24.0
# protobuf==4.25.4
# pyarrow==17.0.0
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pydantic==2.8.2
# pydantic_core==2.20.1
# pydeck==0.9.1
# Pygments==2.18.0
# PyMuPDF==1.24.9
# PyMuPDFb==1.24.9
# pyparsing==3.1.2
# PyPDF2==3.0.1
# pytesseract==0.3.10
# python-dateutil==2.9.0.post0
# pytz==2024.1
# referencing==0.35.1
# requests==2.32.3
# rich==13.7.1
# rpds-py==0.19.1
# rsa==4.9
# six==1.16.0
# smmap==5.0.1
# sniffio==1.3.1
# streamlit==1.32.2
# tenacity==8.5.0
# toml==0.10.2
# toolz==0.12.1
# tornado==6.4.1
# tqdm==4.66.4
# typing_extensions==4.12.2
# tzdata==2024.1
# uritemplate==4.1.1
# urllib3==2.2.2
# watchdog==4.0.1
# yarl==1.9.4
